src/requirements.txt
## Data Processing
pandas
numpy
matplotlib
seaborn

## LLM
openai

## Data
datasets

## AI
dspy-ai

## Environment Variables
python-dotenv